+++ interpreter/emulator/hack/__init__.py
@@ -1,0 +1,1 @@
+"""Batocera/Knulli-oriented HACK emulator launcher assets."""